Charan (#25)
* 11/04/2025

* 22/04/2025 - Charan

* Delete src/main/java/com/realestate/main/security/CorsFilter.java

* 24/04/2025-charan

* 24/04/2025-charan

* 28/04/2025-charan

* 07/05/2025-charan

* 09/05/2025

* 13/05/2025-charan

* 16/05/2025-charan

* 27/05/2025 - charan

* 27/05/2025

* 27/05/2025

* 03/06/2025

* 10/06/2025

* 06/06/2025

* 30/06/2025

+++ src/main/java/com/realestate/main/entity/Venture.java
@@ -33,10 +33,10 @@ public class Venture implements Serializable{
 
 	private String ventureName; 
 	private String ventureSize;
-	private int totalPlots;
-	private int availablePlots;
-	private int bookedPlots;
-	private int soldPlots;
+	private long totalPlots;
+	private long availablePlots;
+	private long bookedPlots;
+	private long soldPlots;
 	private String address;
 	private String city;
 	private String state;
